+++ demo/src/app/components/select/children-demo.ts
@@ -181,17 +181,11 @@ export class ChildrenDemoComponent {
       ]
     }
   ];
-  private value:any = {};
-  private _disabledV:string = '0';
-  private disabled:boolean = false;
+  public value:any = {};
+  public disabled:boolean = false;
 
-  private get disabledV():string {
-    return this._disabledV;
-  }
-
-  private set disabledV(value:string) {
-    this._disabledV = value;
-    this.disabled = this._disabledV === '1';
+  public toggleSelectDisabledState() {
+    this.disabled = !this.disabled;
   }
 
   public selected(value:any):void {